confluence-to-notion/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "confluence-to-notion"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.4",
    "atlassian-python-api>=3.41",
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "pandas>=2.2",
    "matplotlib>=3.9",
    "japanize-matplotlib>=1.1",
    "notion-client>=2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-dotenv>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/confluence_to_notion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["e2e: end-to-end tests requiring live Confluence and Notion APIs"]
env_files = [".env"]
